+++ src/DependencyInjection/Configuration.php
@@ -2,19 +2,20 @@
 
 namespace League\Tactician\Bundle\DependencyInjection;
 
+use League\Tactician\Bundle\DependencyInjection\Compiler\BusBuilder\BusBuildersFromConfig;
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
 use Symfony\Component\Config\Definition\Exception\InvalidConfigurationException;
 
 /**
- * This is the class that validates and merges configuration from your app/config files
+ * This is the class that validates and merges configuration from your app/config files.
  *
  * To learn more see {@link http://symfony.com/doc/current/cookbook/bundles/extension.html#cookbook-bundles-extension-config-class}
  */
 class Configuration implements ConfigurationInterface
 {
     /**
-     * Create a rootnode tree for configuration that can be injected into the DI container
+     * Create a rootnode tree for configuration that can be injected into the DI container.
      *
      * @return TreeBuilder
      */
@@ -38,8 +39,9 @@ class Configuration implements ConfigurationInterface
                                 ->validate()
                                     ->ifTrue(function ($config) {
                                         $isPresent = in_array('tactician.middleware.command_handler', $config);
-                                        $isLast    = end($config) == 'tactician.middleware.command_handler';
-                                        return ($isPresent && !$isLast);
+                                        $isLast = end($config) == 'tactician.middleware.command_handler';
+
+                                        return $isPresent && !$isLast;
                                     })
                                     ->thenInvalid(
                                         '"tactician.middleware.command_handler" should be the last middleware loaded '.
@@ -52,11 +54,11 @@ class Configuration implements ConfigurationInterface
                     ->end()
                 ->end()
                 ->scalarNode('default_bus')
-                    ->defaultValue('default')
+                    ->defaultValue(BusBuildersFromConfig::DEFAULT_BUS_ID)
                     ->cannotBeEmpty()
                 ->end()
                 ->scalarNode('method_inflector')
-                    ->defaultValue('tactician.handler.method_name_inflector.handle')
+                    ->defaultValue(BusBuildersFromConfig::DEFAULT_METHOD_INFLECTOR)
                     ->cannotBeEmpty()
                 ->end()
                 ->arrayNode('security')
@@ -68,13 +70,13 @@ class Configuration implements ConfigurationInterface
                 ->end()
             ->end()
             ->validate()
-                ->ifTrue(function($config) {
+                ->ifTrue(function ($config) {
                     return is_array($config) &&
                         array_key_exists('default_bus', $config) &&
                         array_key_exists('commandbus', $config)
                     ;
                 })
-                    ->then(function($config) {
+                    ->then(function ($config) {
                         $busNames = [];
                         foreach ($config['commandbus'] as $busName => $busConfig) {
                             $busNames[] = $busName;
@@ -89,12 +91,12 @@ class Configuration implements ConfigurationInterface
                                 )
                             );
                         }
+
                         return $config;
                     })
             ->end()
         ;
 
         return $treeBuilder;
-
     }
 }