+++ tests/Fake/OtherFakeCommand.php
@@ -1,0 +1,8 @@
+<?php
+declare(strict_types=1);
+
+namespace League\Tactician\Bundle\Tests\Fake;
+
+final class OtherFakeCommand
+{
+}